+++ R/mainQC.R
@@ -1,6 +1,155 @@
+################################
+###### get and save SAMPLE_QC_summary ####
+#config <- conf; sample_basename <- "DA01";
+getSampleQCSummary <- function(sample_basename, config, save = T){
+  
+  config_tools <- read.csv(file.path(config$tools_path, config$tools_config), stringsAsFactors = FALSE)
+  sampleQC <- list()
+  sampleQC$Sample_ID <- sample_basename
+  
+  ### trimmomatic
+  trimmomatic_out_logfile <- file.path(config$results_path,"logs",paste0(sample_basename,"_",config_tools[config_tools$tool=="trimmomatic","logfile"]))
+  if(file.exists(trimmomatic_out_logfile)){
+    trim_log <- readLines(trimmomatic_out_logfile)
+    trim_log <- str_split(trim_log[startsWith(trim_log, "Input Read Pairs")], " ")
+    sampleQC$Input_read_pairs <- as.double(trim_log[[1]][4])
+    sampleQC$Read_Pairs_Surviving_trimming <- as.double(trim_log[[1]][7])
+    sampleQC$Prc_Read_Pairs_Surviving_trimming <- 100 * sampleQC$Read_Pairs_Surviving_trimming/sampleQC$Input_read_pairs
+  }else{
+    print(paste0("File: ", trimmomatic_out_logfile, " does not exist!"))
+  }
+  
+  ### top_rmdups
+  rmdups_result_file <- file.path(config$results_path, config_tools[config_tools$proces=="mark_duplicates_top", "temp_results_dirs"], sample_basename)
+  top_rmdups_result_file <- file.path(paste0(rmdups_result_file,".top.rmdups_metrics.txt"))
+  
+  if(file.exists(top_rmdups_result_file)){
+    top_dup <- readLines(top_rmdups_result_file)
+    top_dup <- makeDataFrame(data = unlist(str_split(top_dup[startsWith(top_dup, "SAMPLE")], "\t")), 
+                             header = tolower(unlist(str_split(top_dup[startsWith(top_dup, "LIBRARY")], "\t"))))
+    sampleQC$Prc_duplicated_reads_top <- as.numeric(top_dup[,"percent_duplication"])*100
+  }else{
+    print(paste0("File: ", top_rmdups_result_file, " does not exist!"))
+  }
+  
+  ### bottom_rmdups
+  bottom_rmdups_result_file <- file.path(paste0(rmdups_result_file,".bottom.rmdups_metrics.txt"))
+  if(file.exists(bottom_rmdups_result_file)){
+    bottom_dup <- readLines(bottom_rmdups_result_file)
+    bottom_dup <- makeDataFrame(data = unlist(str_split(bottom_dup[startsWith(bottom_dup, "SAMPLE")], "\t")), 
+                                header = tolower(unlist(str_split(bottom_dup[startsWith(bottom_dup, "LIBRARY")], "\t"))))
+    sampleQC$Prc_duplicated_reads_bottom <- as.numeric(bottom_dup[,"percent_duplication"])*100
+  }else{
+    print(paste0("File: ", bottom_rmdups_result_file, " does not exist!"))
+  }
+  
+  ### basic_mapping
+  basic_mapping_metrics_result_file <- file.path(config$results_path, config_tools[config_tools$proces=="basic_mapping_metrics","temp_results_dirs"], sample_basename)
+  basic_mapping_metrics_result_file <- paste0(basic_mapping_metrics_result_file,"_basic_mapping_metrics.txt")
+  if(file.exists(basic_mapping_metrics_result_file)){
+    metrics <- readLines(basic_mapping_metrics_result_file)
+    headerStartLine <- which(startsWith(metrics,"CATEGORY"))
+    metrics_df <- makeDataFrame(data = unlist(str_split(metrics[headerStartLine + 1], "\t")), 
+                                header = tolower(unlist(str_split(metrics[headerStartLine], "\t"))))
+    metrics_df <- rbind(metrics_df, unlist(str_split(metrics[headerStartLine + 2], "\t")))
+    metrics_df <- rbind(metrics_df, unlist(str_split(metrics[headerStartLine + 3], "\t")))
+    sampleQC$Number_of_reads_after_removing_duplicates <- metrics_df[metrics_df$category == "PAIR", "pf_reads"]
+  }else{
+    print(paste0("File: ", basic_mapping_metrics_result_file, " does not exist!"))
+  }
+  
+  ### flagstat
+  flagstat_result_file <- file.path(file.path(config$results_path,"logs"), paste0(sample_basename,"_",config_tools[config_tools$process=="flagstat_filtered_bam","logfile"]))
+  if(file.exists(flagstat_result_file)){
+    metrics <- readLines(flagstat_result_file)
+    metrics <- unlist(str_split(metrics[1]," "))
+    sampleQC$Number_of_reads_after_filtering <- as.numeric(metrics[1])
+    sampleQC$Prc_passed_filtering_step <- 100 * as.numeric(sampleQC$Number_of_reads_after_filtering)/as.numeric(sampleQC$Number_of_reads_after_removing_duplicates)
+  }else{
+    print(paste0("File: ", flagstat_result_file, " does not exist!"))
+  }
+  
+  ### on_target_reads
+  on_target_result_file <- file.path(config$results_path, config_tools[config_tools$proces=="on_target_reads","temp_results_dirs"], sample_basename)
+  on_target_result_file <- paste0(on_target_result_file,"_on_target_reads.txt")
+  if(file.exists(on_target_result_file)){
+    sampleQC$Number_of_on_target_reads <- as.numeric(yaml::yaml.load_file(on_target_result_file)[['number_of_on_target_reads']])
+    sampleQC$Prc_of_on_target_reads <- 100 * sampleQC$Number_of_on_target_read/as.numeric(sampleQC$Number_of_reads_after_removing_duplicates)
+  }else{
+    print(paste0("File: ", on_target_result_file, " does not exist!"))
+  }
+  
+  ### depth_of_cov
+  depth_of_cov_result_file <- file.path(config$results_path, config_tools[config_tools$proces=="depth_of_coverage","temp_results_dirs"], sample_basename)
+  depth_of_cov_result_file <- paste0(depth_of_cov_result_file,"_gatk_target_coverage.sample_summary")
+  if(file.exists(depth_of_cov_result_file)){
+    depth_summary <- readLines(depth_of_cov_result_file)
+    depth_summary <- makeDataFrame(data = unlist(str_split(depth_summary[2], "\t")), 
+                                   header = tolower(unlist(str_split(depth_summary[1], "\t"))))
+    sampleQC$Mean_coverage <- as.numeric(depth_summary[1, "mean"])
+  }else{
+    print(paste0("File: ", on_target_result_file, " does not exist!"))
+  }
+  
+  ### Calculate conversion efficiency
+  methyl_result_file <- file.path(config$results_path, config_tools[config_tools$proces=="methratio","temp_results_dirs"], sample_basename)
+  methyl_result_data <- readMethylResultData(paste0(methyl_result_file,".methylation_results.bed.panel"))
+  methyl_result_data <- data.frame(methyl_result_data)
+  if(!is.null(methyl_result_data)){
+    control <- methyl_result_data[methyl_result_data$chr == config$ref_control_sequence_name,]
+    sampleQC$Number_of_Cs_in_control <- nrow(control)
+    conversion_eff <- c(100 * (1-sum(control$C_count)/sum(control$CT_count)))
+    sampleQC$Conversion_eff <- conversion_eff
+    
+    #Add to QC Sample report
+    methyl_res_panel_df_no_control <- methyl_result_data[methyl_result_data$chr != config$ref_control_sequence_name,]
+    sampleQC$Number_of_Cs_in_panel <- nrow(methyl_res_panel_df_no_control)
+    
+    methyl_res_panel_df_no_control_CpG <- methyl_res_panel_df_no_control[methyl_res_panel_df_no_control$context=='CG',]
+    sampleQC$Number_of_Cs_in_panel_CpG <- nrow(methyl_res_panel_df_no_control_CpG)
+    
+    methyl_res_panel_df_no_control_nonCpG <- methyl_res_panel_df_no_control[methyl_res_panel_df_no_control$context!='CG',]
+    sampleQC$Number_of_Cs_in_panel_non_CpG <- nrow(methyl_res_panel_df_no_control_nonCpG)
+    
+    methylSplitResult <- getMethylSplitByCT_Count(methyl_result_data, config$ref_control_sequence_name, 10)
+    sampleQC$Number_of_Cs_in_panel_CpG_cov_min10 <- nrow(methylSplitResult$methyl_res_panel_df_no_control_CpG)
+    sampleQC$Number_of_Cs_in_panel_non_CpG_cov_min10 <- nrow(methylSplitResult$methyl_res_panel_df_no_control_nonCpG)
+    
+    #methyl_res_panel_df_no_control_CpG_max9 <- methyl_res_panel_df_no_control_CpG[methyl_res_panel_df_no_control_CpG$eff_CT_count<10,]
+    sampleQC$Number_of_Cs_in_panel_CpG_cov_max9 <- sum(methyl_res_panel_df_no_control_CpG$eff_CT_count<10)
+    
+    sampleQC$Prc_of_Cs_in_panel_CpG_cov_min10 <- 100 * (sampleQC$Number_of_Cs_in_panel_CpG_cov_min10/sampleQC$Number_of_Cs_in_panel_CpG)
+    sampleQC$Prc_of_Cs_in_panel_CpG_cov_max9 <- 100 * (sampleQC$Number_of_Cs_in_panel_CpG_cov_max9/sampleQC$Number_of_Cs_in_panel_CpG)
+  }
+  
+  if(save){
+    # Write QC report to yaml file
+    sample_report_file <- file.path(config$results_path, "QC_report", paste0(sample_basename,"_QC_summary.yml"))
+    print(paste0("Saving QC report file to: ", sample_report_file))
+    yaml::write_yaml(lapply(sampleQC,as.character), sample_report_file)
+  }
+  
+  return(sampleQC)
+}
+
 ####################################################
 ###### Table of QC common for all present samples ####
-getQCSummary <- function(config, save = T){
+createAllQCSummary <- function(config){
+  config_tools <- read.csv(file.path(config$tools_path, config$tools_config), stringsAsFactors = FALSE)
+  samples <- list.files(file.path(config$results_path,"logs"),full.names = F)
+  samples <- samples[endsWith(samples, config_tools[config_tools$tool=="trimmomatic","logfile"])]
+  samples <- str_replace_all(samples, pattern = paste0("_",config_tools[config_tools$tool=="trimmomatic","logfile"]),replace = "")
+  
+  for(s in samples){
+    print(paste0("Processing sample: ", s))
+          sampleQC <- getSampleQCSummary(s, config)
+  }
+  return(T)
+}
+
+####################################################
+###### Table of QC common for all present samples ####
+readAllQCSummary <- function(config, save = T){
   
   input_dir <- file.path(config$results_path, "QC_report")
   input_files <- list.files(input_dir, pattern = "\\_QC_summary.yml$", full.names = T)
@@ -8,8 +157,15 @@ getQCSummary <- function(config, save = T){
   if(length(input_files) == 0)
     stop(paste0("There is no sample result files to create the report. Directory", input_dir))
   
-  qc_summary <- lapply(input_files, function(x) as.data.frame(yaml::read_yaml(x)))
+  qc_summary <- lapply(input_files, function(x) as.data.frame(yaml::read_yaml(x), stringsAsFactors = F))
   qc_summary <- data.frame(rbindlist(qc_summary))
+  
+  no_numeric_cols <- c("Sample_ID")
+  for(curcol in names(qc_summary)){
+    if(!curcol %in% no_numeric_cols){
+      qc_summary[,curcol] <- as.numeric(qc_summary[,curcol])
+    }
+  }
   
   if(save){
     fwrite(qc_summary, file = file.path(config$results_path,"QC_report/","SummaryQC.csv"))
@@ -19,7 +175,7 @@ getQCSummary <- function(config, save = T){
 
 ####################################################
 #### PLOT: Number of sites in CpG context covered by more/less than 10
-plotSitesCovBy10 <- function(qc_summary, config, pal = brewer.pal(8, "Dark2"), share = F, save = T, fontsize = 14){
+plotSitesCovBy10 <- function(qc_summary, config, pal = brewer.pal(8, "Dark2"), share = F, save = T, fontsize = 10){
   
   qc_summary_coverage <- rbind(data.frame(SampleID = qc_summary$Sample_ID, Number = qc_summary$Number_of_Cs_in_panel_CpG_cov_min10, Coverage = "cov>=10", stringsAsFactors = F),
                                data.frame(SampleID = qc_summary$Sample_ID, Number = qc_summary$Number_of_Cs_in_panel_CpG_cov_max9, Coverage = "cov<10", stringsAsFactors = F))
@@ -39,7 +195,7 @@ plotSitesCovBy10 <- function(qc_summary, config, pal = brewer.pal(8, "Dark2"), s
     theme(legend.text = element_text(size = fontsize)) + 
     ylab("Number of sites") +
     xlab("Sample ID") +
-    ggtitle("Number of sites in CpG context covered by more/less than 10.")
+    ggtitle("Number of sites in CpG context \ncovered by more/less than 10")
   
   if(save){
     ggsave(filename = file.path(config$results_path,"QC_report/",paste0("SummarySitesCovBy10.",config$plot_format)), plot = gg, device=config$plot_format)
@@ -50,7 +206,7 @@ plotSitesCovBy10 <- function(qc_summary, config, pal = brewer.pal(8, "Dark2"), s
 
 ####################################################
 #### PLOT: Number of sites covered by minimum 10 reads \nseparately for sites in CpG and non-CpG context
-plotSitesCovBy10CpGnonCpG <- function(qc_summary, config, pal = brewer.pal(8, "Dark2"), share = F, save = T, fontsize = 14){
+plotSitesCovBy10CpGnonCpG <- function(qc_summary, config, pal = brewer.pal(8, "Dark2"), share = F, save = T, fontsize = 10){
   
   qc_summary_cpg <- rbind(data.frame(SampleID = qc_summary$Sample_ID, Number = qc_summary$Number_of_Cs_in_panel_CpG_cov_min10, Context = "CpG", stringsAsFactors = F),
                           data.frame(SampleID = qc_summary$Sample_ID, Number = qc_summary$Number_of_Cs_in_panel_non_CpG_cov_min10, Context = "non-CpG", stringsAsFactors = F))
@@ -69,7 +225,7 @@ plotSitesCovBy10CpGnonCpG <- function(qc_summary, config, pal = brewer.pal(8, "D
     theme(legend.text = element_text(size = fontsize)) + 
     ylab("Number of sites") +
     xlab("Sample ID") +
-    ggtitle("Number of sites covered by minimum 10 reads \nseparately for sites in CpG and non-CpG context.")
+    ggtitle("Number of cytosines covered by at least 10 reads \nin the CpG and non-CpG context")
   if(save){
     ggsave(filename=file.path(config$results_path,"QC_report/",paste0("SummarySitesCovBy10CpGnonCpG.",config$plot_format)), plot=gg, device=config$plot_format)
   }
@@ -79,21 +235,31 @@ plotSitesCovBy10CpGnonCpG <- function(qc_summary, config, pal = brewer.pal(8, "D
 
 ##############################################
 ######## Read data using MethylKit ############
-readMethData <- function(config){
+readMethData <- function(config, context = c('CpG', 'non-CpG')){
   
   result_dir <- file.path(config$results_path, "methyl_results")
-  result_files <- list.files(result_dir, pattern = "\\.CpG_min10", full.names = T)
+  if(context[1] == 'CpG'){
+    file_sufix <- "CpG_min10"
+    methcontext <- "CpG"
+  }else{
+    file_sufix <- "non_CpG_min10"
+    methcontext <- c("CpH", "CHH", "CHG")
+  }
+  
+  result_files <- list.files(result_dir, pattern = paste0("\\.", file_sufix), full.names = T)
   sample_id <- lapply(strsplit(basename(result_files), '\\.'), function(x){x[1]})
   
-  methData <- methRead(as.list(result_files),
-                       sample.id = sample_id,
-                       treatment = c(rep(0, length(sample_id))), # 0/1 control/test samples
-                       assembly = "hg38",
-                       header = TRUE,
-                       context = "CpG",
-                       resolution = "base",
-                       pipeline = list(fraction=TRUE, chr.col=1, start.col=2, end.col=3, coverage.col=7, strand.col=4, freqC.col=6)
-  )
+  methData <- methylKit::methRead(as.list(result_files),
+                                  sample.id = sample_id,
+                                  treatment = c(rep(0, length(sample_id))), # 0/1 control/test samples
+                                  assembly = "hg38",
+                                  header = TRUE,
+                                  context = methcontext,
+                                  resolution = "base",
+                                  pipeline = list(fraction=TRUE, chr.col=1, start.col=2, end.col=3, coverage.col=7, strand.col=4, freqC.col=6))
+  
+  attr(methData, 'context') <- context
+  
   return(methData)
 }
 
@@ -114,7 +280,7 @@ plotSingleMethStats <- function(single_meth_data){
 ###### Generate and save Basic MethylKit stats for all samples #######
 plotMethStats <- function(meth_data, config, pal = brewer.pal(8, "Dark2"), save = T){
   for(i in 1:length(meth_data)){
-    openPlotFile(file.path(config$results_path,'QC_report',paste0(meth_data[[i]]@sample.id,"_histCpGStats.",config$plot_format)))
+    openPlotFile(file.path(config$results_path, 'QC_report', paste0(meth_data[[i]]@sample.id,"_histCpGStats.", config$plot_format)))
     plotSingleMethStats(meth_data[[i]])
     dev.off()
   }
@@ -122,15 +288,23 @@ plotMethStats <- function(meth_data, config, pal = brewer.pal(8, "Dark2"), save 
 }
 
 ####################################################
-###### PLOT: Generate and save log10 Coverage boxplot for all samples
-plotMethStatsSummary <- function(meth_data, config, pal = brewer.pal(8, "Dark2"), save = T, fontsize = 14){
+###### PLOT: Generate and save Coverage boxplot for all samples (by default log10 coverage)
+plotMethStatsSummary <- function(meth_data, config, log = TRUE, pal = brewer.pal(8, "Dark2"), save = T, fontsize = 10){
   
   sampleCov <- lapply(meth_data, function(x){
-    data.frame(coverage = log10(x$coverage), sample_id = x@sample.id)
+    data.frame(coverage = x$coverage, sample_id = x@sample.id)
   })
+  
   sampleCov <- rbindlist(sampleCov)
   sampleCov$sample_id <- factor(sampleCov$sample_id)
   
+  xlab <- "Read coverage per base"
+  outfile_sufix <- ""
+  if(log){
+    sampleCov$coverage = log10(sampleCov$coverage)
+    xlab <- "Read coverage per base [log10]"
+    outfile_sufix <- "Log10"
+  }
   gg <- ggplot(sampleCov, aes(x=sample_id, y=coverage)) + 
     geom_boxplot(outlier.colour="darkred", outlier.shape=1, outlier.size=0.5, fill = pal[1], color="black") +
     theme_minimal() +
@@ -139,12 +313,12 @@ plotMethStatsSummary <- function(meth_data, config, pal = brewer.pal(8, "Dark2")
     theme(axis.text.y = element_text(hjust = 1, size = fontsize)) +
     theme(legend.text=element_text(size = fontsize)) + 
     #theme(plot.title = element_text(hjust = 0.5, size = 14)) +
-    ggtitle("CpG Coverage for the experiment") +
-    scale_y_continuous(name = "log 10 of read Coverage per base") +
-    scale_x_discrete(name = "Sample Id") 
+    ggtitle("CpG coverage for the experiment") +
+    scale_y_continuous(name = xlab) +
+    scale_x_discrete(name = "Sample ID") 
   
   if(save){
-    plotfile <- file.path(config$results_path,'QC_report',paste0('SummaryCpGCoverage.',config$plot_format))
+    plotfile <- file.path(config$results_path,'QC_report',paste0('SummaryCpGCoverage',outfile_sufix,".",config$plot_format))
     ggsave(plotfile, gg)
   }
   
@@ -153,20 +327,29 @@ plotMethStatsSummary <- function(meth_data, config, pal = brewer.pal(8, "Dark2")
 
 ####################################################
 ### PLOT: Beta Values Boxplot
-plotBetaValuesSummary <- function(meth_data, config, pal = brewer.pal(8, "Dark2"), save = T, fontsize = 14){
+#meth_data <- methData
+#config <- conf
+plotBetaValuesSummary <- function(meth_data, config, sample_size = 100000, pal = brewer.pal(8, "Dark2"), save = T, fontsize = 10){
+  
+  data_context <- attr(meth_data, 'context')
+  plot_title <- paste0("Beta values of ", data_context, " sites across samples")
   
   percentage_meth <- lapply(meth_data, function(x){x$numCs/x$coverage } )
   methyl_levels <- list()
+  i =1
   for (i in 1:length(meth_data)){
     methyl_levels[[i]] <- data.frame(sample.id = meth_data[[i]]@sample.id, 
                                      beta_values = as.numeric(percentage_meth[[i]]),
                                      stringsAsFactors = F)
+    if(sample_size > 0 & !is.na(sample_size) & sample_size < nrow(methyl_levels[[i]])){
+      methyl_levels[[i]] <- (methyl_levels[[i]])[sample(nrow(methyl_levels[[i]]), sample_size), ]
+    }
   }
   methyl_levels <- rbindlist(methyl_levels)
   methyl_levels <- methyl_levels[order(methyl_levels$sample.id),]
   methyl_levels <- data.frame(methyl_levels)
   methyl_levels$sample.id <- factor(methyl_levels$sample.id, levels = sort(unique(methyl_levels$sample.id)))
-  
+
   gg <- ggplot(methyl_levels, aes(x=sample.id, y=beta_values)) +
     geom_boxplot(outlier.colour="red", outlier.shape=42, outlier.size=3, notch=FALSE, fill=pal[1], color="black") +
     theme_minimal() +
@@ -174,12 +357,12 @@ plotBetaValuesSummary <- function(meth_data, config, pal = brewer.pal(8, "Dark2"
     theme(axis.text.x = element_text(angle = 60, hjust = 1, size = fontsize)) +
     theme(axis.text.y = element_text(hjust = 1, size = fontsize)) +
     theme(legend.position="none") +  
-    ggtitle("Methylation over samples ") +
-    xlab("Sample Id") + 
-    ylab("Beta Value")
+    ggtitle(plot_title) +
+    xlab("Sample ID") + 
+    ylab("Beta value")
   
   if(save){
-    plotfile <- file.path(config$results_path,'QC_report',paste0('BetaValuesSummary.',config$plot_format))
+    plotfile <- file.path(config$results_path,'QC_report', paste0('BetaValuesSummary', data_context, '.', config$plot_format))
     ggsave(plotfile, gg)
   }
   
@@ -187,7 +370,7 @@ plotBetaValuesSummary <- function(meth_data, config, pal = brewer.pal(8, "Dark2"
 }
 ####################################################
 ### PLOT: Methylation Levels
-plotMethLevels <- function(meth_data, config, breaks = c(0,0.1,0.2,0.4,0.6,0.8,0.9,1), pal = brewer.pal(8, "Dark2"), share = F, save = T, fontsize = 14){
+plotMethLevels <- function(meth_data, config, breaks = c(0,0.1,0.2,0.4,0.6,0.8,0.9,1), pal = brewer.pal(8, "Dark2"), share = F, save = T, fontsize = 10){
   
   percentage_meth <- lapply(meth_data, function(x){x$numCs/x$coverage } )
   methyl_levels <- list()
@@ -218,8 +401,8 @@ plotMethLevels <- function(meth_data, config, breaks = c(0,0.1,0.2,0.4,0.6,0.8,0
     theme(axis.text.x = element_text(angle = 60, hjust = 1, size = fontsize)) +
     theme(axis.text.y = element_text(hjust = 1, size = fontsize)) +
     theme(legend.text=element_text(size = fontsize)) + 
-    ggtitle("Methylation Levels") +
-    xlab("Sample Id") + 
+    ggtitle("Frequency of CpGs according to \nspecific methylation level ranges") +
+    xlab("Sample ID") + 
     ylab("Frequency")
   
   if(save){
@@ -234,7 +417,7 @@ plotMethLevels <- function(meth_data, config, breaks = c(0,0.1,0.2,0.4,0.6,0.8,0
 ###### PLOT: Hyper/Hypo CpG genom & islands annotation for each sample.
 #meth_data <- methData
 #config <- conf
-plotCpGAnnotation <- function(meth_data, hypo_hyper_def = c(0.2,0.8), config, pal = brewer.pal(8, "Dark2"), share = F, save = T, fontsize = 14){
+plotCpGAnnotation <- function(meth_data, hypo_hyper_def = c(0.2,0.8), config, pal = brewer.pal(8, "Dark2"), share = F, save = T, fontsize = 10){
   
   cpg.gene <- readTranscriptFeatures(file.path(config$ref_data_path, config$ref_data_CpGGenomAnnotation))
   cpg.island <- readFeatureFlank(file.path(config$ref_data_path, config$ref_data_CpgIslandAnnotation), feature.flank.name=c("CpGi","shores"))
@@ -258,7 +441,7 @@ plotCpGAnnotation <- function(meth_data, hypo_hyper_def = c(0.2,0.8), config, pa
 
 ####################################################
 ###### PLOT: Hyper/Hypo CpG genom annotation for each sample.
-plotCpGGenomAnnotation <- function(meth_data, gene_annot_data, config, pal = brewer.pal(8, "Dark2"), subtitle = "", share = F, save = T, fontsize = 14){
+plotCpGGenomAnnotation <- function(meth_data, gene_annot_data, config, pal = brewer.pal(8, "Dark2"), subtitle = "", share = F, save = T, fontsize = 10){
   
   annot_summary <- lapply(meth_data, function(x){ 
     annot <- annotateWithGeneParts(as(x,"GRanges"), gene_annot_data) 
@@ -284,8 +467,8 @@ plotCpGGenomAnnotation <- function(meth_data, gene_annot_data, config, pal = bre
     theme(axis.text.x = element_text(angle = 60, hjust = 1, size = fontsize)) +
     theme(axis.text.y = element_text(hjust = 1, size = fontsize)) +
     theme(legend.text=element_text(size = fontsize)) + 
-    ggtitle(paste(subtitle,"CpG Genom Annotation")) +
-    xlab("Sample Id") + 
+    ggtitle(paste(subtitle,"CpGs annotated to \ngenomic regions")) +
+    xlab("Sample ID") + 
     ylab("Frequency")
   
   if(save){
@@ -297,7 +480,7 @@ plotCpGGenomAnnotation <- function(meth_data, gene_annot_data, config, pal = bre
 
 ####################################################
 ###### PLOT: Hyper/Hypo CpG islands annotation for each sample.
-plotCpGIslandsAnnotation <- function(meth_data, cpg_annot_data, config, pal = brewer.pal(8, "Dark2"), subtitle = "", share = F, save = T, fontsize = 14){
+plotCpGIslandsAnnotation <- function(meth_data, cpg_annot_data, config, pal = brewer.pal(8, "Dark2"), subtitle = "", share = F, save = T, fontsize = 10){
   
   annot_summary <- lapply(meth_data, function(x){ 
     annot <- annotateWithFeatureFlank(as(x,"GRanges"), cpg_annot_data$CpGi, cpg_annot_data$shores, feature.name="CpGi", flank.name="shores")
@@ -323,8 +506,8 @@ plotCpGIslandsAnnotation <- function(meth_data, cpg_annot_data, config, pal = br
     theme(axis.text.x = element_text(angle = 60, hjust = 1, size = fontsize)) +
     theme(axis.text.y = element_text(hjust = 1, size = fontsize)) +
     theme(legend.text=element_text(size = fontsize)) + 
-    ggtitle(paste(subtitle,"CpG Islands Annotation")) +
-    xlab("Sample Id") + 
+    ggtitle(paste(subtitle,"CpGs annotated to \nCpG-islands")) +
+    xlab("Sample ID") + 
     ylab("Frequency")
   
   if(save){
@@ -336,7 +519,7 @@ plotCpGIslandsAnnotation <- function(meth_data, cpg_annot_data, config, pal = br
 
 ###################################################
 ##### Number of Common CpG Shared Between Samples
-plotCntCommonCpG <- function(meth_data, config, pal = brewer.pal(8, "Dark2"), save = T, fontsize = 14){
+plotCntCommonCpG <- function(meth_data, config, pal = brewer.pal(8, "Dark2"), save = T, fontsize = 10){
   
   commonCpg <- lapply(meth_data, function(x){
     data.frame(cpg = paste0(x$chr, "_", x$start))
@@ -351,7 +534,7 @@ plotCntCommonCpG <- function(meth_data, config, pal = brewer.pal(8, "Dark2"), sa
     theme(axis.text.x = element_text(angle = 60, hjust = 1, size = fontsize)) +
     theme(axis.text.y = element_text(hjust = 1, size = fontsize)) +
     theme(legend.text=element_text(size = fontsize)) + 
-    ggtitle("Common CpG Shared Between Samples") +
+    ggtitle("Common CpG shared between samples") +
     xlab("Number of samples") + 
     ylab("Number of common CpG")